feat: adds derived resources and moves highwater from postprocessing (#176)
* feat: adds derived resources and moves highwater from postprocessing

* fix: fixes in derived resources

* fix: style issues

* fix: Fix typing

* tests: remove unnecessary test

* fix: fixed compatibility with python 3.10

* fix: style issues

---------
