+++ tests/compilation/test_postprocessing.py
@@ -13,27 +13,12 @@
 # limitations under the License.
 
 from dataclasses import replace
-from pathlib import Path
 
-import pytest
-import yaml
-from qref.schema_v1 import RoutineV1, SchemaV1
+from qref.schema_v1 import RoutineV1
 
 from bartiq import compile_routine
 from bartiq._routine import Routine
-from bartiq.compilation.postprocessing import add_qubit_highwater, aggregate_resources
-
-
-def load_highwater_test_data():
-    test_file = Path(__file__).parent / "data/highwater.yaml"
-    data = []
-    with open(test_file) as f:
-        for original, expected in yaml.safe_load(f):
-            data.append((SchemaV1(**original), SchemaV1(**expected)))
-    return data
-
-
-HIGHWATER_TEST_DATA = load_highwater_test_data()
+from bartiq.compilation.postprocessing import aggregate_resources
 
 
 def _get_simple_routine(backend):
@@ -90,90 +75,3 @@ def test_aggregate_resources(backend):
     compiled_routine = compile_routine(routine, postprocessing_stages=postprocessing_stages, backend=backend).routine
     assert len(compiled_routine.resources) == 1
     assert compiled_routine.resources["op"].value == 22
-
-
-@pytest.mark.parametrize("routine, expected_routine", HIGHWATER_TEST_DATA)
-def test_add_qubit_highwater(routine, expected_routine, backend):
-    postprocessing_stages = [add_qubit_highwater]
-    compiled_routine = compile_routine(routine, postprocessing_stages=postprocessing_stages, backend=backend)
-    assert compiled_routine.to_qref() == expected_routine
-
-
-def test_add_qubit_highwater_with_custom_names(backend):
-    input_routine = SchemaV1(
-        **{
-            "version": "v1",
-            "program": {
-                "name": "root",
-                "type": None,
-                "ports": [
-                    {"name": "in_0", "direction": "input", "size": "N"},
-                    {"name": "out_0", "direction": "output", "size": "N"},
-                ],
-                "resources": [
-                    {"name": "local_ancillae", "type": "qubits", "value": 7},
-                    {"name": "custom_ancillae", "type": "qubits", "value": 5},
-                    {"name": "qubit_highwater", "type": "qubits", "value": "N+7"},
-                ],
-            },
-        }
-    )
-
-    def custom_stage(routine, backend):
-        return add_qubit_highwater(routine, backend, resource_name="custom_highwater", ancillae_name="custom_ancillae")
-
-    postprocessing_stages = [custom_stage]
-    compiled_routine = compile_routine(
-        input_routine, postprocessing_stages=postprocessing_stages, backend=backend
-    ).routine
-
-    assert str(compiled_routine.resources["custom_highwater"].value) == "N + 5"
-
-
-def test_computing_highwater_for_non_chronologically_sorted_routine_raises_warning():
-    input_routine = {
-        "version": "v1",
-        "program": {
-            "name": "root",
-            "type": None,
-            "ports": [
-                {"name": "in_0", "direction": "input", "size": "N"},
-                {"name": "out_0", "direction": "output", "size": "N"},
-            ],
-            "children": [
-                {
-                    "name": "a",
-                    "type": "a",
-                    "ports": [
-                        {"name": "in_0", "direction": "input", "size": "K"},
-                        {"name": "out_0", "direction": "output", "size": "K"},
-                    ],
-                },
-                {
-                    "name": "b",
-                    "type": "b",
-                    "ports": [
-                        {"name": "in_0", "direction": "input", "size": "K"},
-                        {"name": "out_0", "direction": "output", "size": "K"},
-                    ],
-                },
-            ],
-            "connections": ["in_0 -> b.in_0", "b.out_0 -> a.in_0", "a.out_0 -> out_0"],
-        },
-    }
-
-    with pytest.warns(
-        match=(
-            "Order of children in provided routine does not match the topology. Bartiq will use one of topological "
-            "orderings as an estimate of chronology, but the computed highwater value might be incorrect."
-        )
-    ):
-        _ = compile_routine(input_routine, postprocessing_stages=[add_qubit_highwater])
-
-
-def test_highwater_of_an_empty_routine_is_zero():
-    input_routine = {"version": "v1", "program": {"name": "root"}}
-
-    compiled_routine = compile_routine(input_routine, postprocessing_stages=[add_qubit_highwater]).routine
-
-    assert compiled_routine.resources["qubit_highwater"].value == 0
